bot/requirements.txt
python-dotenv
aiogram
redis
psycopg2
confluent_kafka